refactor(cli): tighten _persist.py to shared kernel, move orchestration path to _orchestration.py
- Move setup_orchestration_persist + register_branch_hook from _persist.py into
  _orchestration.py where they're defined and called; nothing outside that module
  ever imported them directly
- Merge teardown_agent_persist and teardown_orchestration_persist into a single
  teardown_persist that handles both hook shapes (ctx["hook"] for agent,
  ctx["hooks"] for orchestration); old names kept as aliases for any external
  callers
- Extract _resolve_project helper to eliminate the duplicated project-detection
  block that both setup paths shared
- Add logging/time/uuid to _orchestration.py top-level imports; remove redundant
  inline imports in finalize_orchestration and resolve_modes
- Result: _persist.py 558 -> 336 lines; _orchestration.py absorbs the moved code
  cleanly; 786 CLI+state tests pass unchanged
