+++ lionagi/cli/orchestrate/_orchestration.py
@@ -5,7 +5,10 @@
 from __future__ import annotations
 
 import json
+import logging
 import os
+import time
+import uuid
 from dataclasses import dataclass, field
 from pathlib import Path
 from typing import Any
@@ -17,13 +20,7 @@ from lionagi.state import provenance as _provenance
 
 from .._agents import AgentProfile, list_agents, load_agent_profile
 from .._logging import hint
-from .._persist import (
-    register_branch_hook as _register_branch_hook,
-)
-from .._persist import (
-    setup_orchestration_persist,
-    teardown_orchestration_persist,
-)
+from .._persist import _resolve_project, teardown_persist
 from .._providers import build_imodel_from_spec, resolve_persisted_effort
 from .._runs import RunDir, allocate_run, save_last_branch_pointer
 
@@ -120,8 +117,6 @@ def role_config(role: str):
 
 def resolve_modes(role: str, override: list[str] | None = None) -> list[str]:
     """Validated cognitive modes for a role: override gated by modes_allow."""
-    import logging
-
     from lionagi.casts.pattern import Mode
 
     cfg = role_config(role)
@@ -447,7 +442,7 @@ def build_worker_branch(
         _grant_spawn(wb)
 
     if env._live_persist:
-        _register_branch_hook(env._live_persist, wb)
+        register_branch_hook(env._live_persist, wb)
 
     return wb, w_model, w_profile
 
@@ -461,9 +456,6 @@ def finalize_orchestration(
     emit_hints: bool = True,
 ) -> tuple[list[tuple[str, str, str]], str]:
     """Persist branch snapshots + last-branch pointer + hints."""
-    import logging
-
-    # Ensure branches_dir exists (allocate_run did, but be defensive).
     env.run.ensure_state_dirs()
     log = logging.getLogger("lionagi.cli")
 
@@ -499,6 +491,192 @@ def finalize_orchestration(
                 hint(f'[{bname}]      li agent -r {bid} "..."')
 
     return branch_ids, orc_branch_id
+
+
+_log_orch = logging.getLogger("lionagi.cli")
+
+
+async def setup_orchestration_persist(
+    session: Any,
+    *,
+    invocation_kind: str | None = None,
+    playbook_name: str | None = None,
+    agent_name: str | None = None,
+    artifacts_path: str | None = None,
+    artifact_contract: dict | None = None,
+    invocation_id: str | None = None,
+    model: str | None = None,
+    provider: str | None = None,
+    effort: str | None = None,
+    project: str | None = None,
+    branches: list[Any] | None = None,
+) -> dict | None:
+    from lionagi.state import provenance as _provenance
+    from lionagi.state.db import StateDB
+
+    db = None
+    try:
+        db = StateDB()
+        await db.open()
+
+        session_id = str(session.id)
+        session_dict = session.to_dict(mode="db")
+
+        session_prog_id = str(uuid.uuid4())
+        await db.create_progression(session_prog_id)
+
+        _proj, _proj_src = _resolve_project(project)
+        from lionagi.cli.kill import current_pid_markers as _pid_markers
+
+        _identity_markers = _pid_markers()
+        _node_meta = {**(session_dict.get("node_metadata") or {}), **_identity_markers}
+        await db.create_session(
+            {
+                "id": session_id,
+                "created_at": session_dict["created_at"],
+                "node_metadata": _node_meta,
+                "name": session_dict.get("name"),
+                "user": session_dict.get("user"),
+                "progression_id": session_prog_id,
+                "first_msg_id": None,
+                "last_msg_id": None,
+                "invocation_kind": invocation_kind,
+                "playbook_name": playbook_name,
+                "agent_name": agent_name,
+                "artifacts_path": artifacts_path,
+                "artifact_contract_json": artifact_contract,
+                "status": "running",
+                "started_at": time.time(),
+                "invocation_id": invocation_id,
+                "model": _provenance.resolve_model_spec(provider, model),
+                "provider": provider,
+                "effort": effort,
+                "agent_hash": _provenance.agent_definition_hash(agent_name),
+                "project": _proj,
+                "project_source": _proj_src,
+            }
+        )
+
+        ctx: dict[str, Any] = {
+            "db": db,
+            "session": session,
+            "session_id": session_id,
+            "session_prog_id": session_prog_id,
+            "branch_prog_ids": {},
+            "hooks": [],
+            "artifacts_path": artifacts_path,
+            "artifact_contract": artifact_contract,
+            "identity_markers": _identity_markers,
+        }
+
+        for branch in branches or []:
+            register_branch_hook(ctx, branch)
+
+        return ctx
+    except Exception as exc:
+        _log_orch.warning(
+            "live persist setup failed (%s) — disabling persistence for this run",
+            exc,
+            exc_info=True,
+        )
+        if db is not None:
+            try:
+                await db.close()
+            except Exception as close_exc:
+                _log_orch.warning(
+                    "fallback db.close after setup failure also failed: %s", close_exc
+                )
+        return None
+
+
+def register_branch_hook(ctx: dict[str, Any], branch: Any) -> None:
+    from lionagi.ln.concurrency import Lock
+
+    db = ctx["db"]
+    session_id = ctx["session_id"]
+    session_prog_id = ctx["session_prog_id"]
+    branch_id = str(branch.id)
+
+    branch_prog_id = str(uuid.uuid4())
+    ctx["branch_prog_ids"][branch_id] = branch_prog_id
+    initialized = {"done": False}
+    init_lock = Lock()
+
+    async def _ensure_branch_row():
+        if initialized["done"]:
+            return
+        async with init_lock:
+            if initialized["done"]:
+                return
+
+            await db.create_progression(branch_prog_id)
+
+            branch_dict = branch.to_dict(mode="db")
+            node_meta = branch_dict.get("node_metadata") or {}
+            if isinstance(node_meta, str):
+                node_meta = json.loads(node_meta)
+            if "chat_model" in branch_dict:
+                node_meta["chat_model"] = branch_dict["chat_model"]
+            node_meta = json.loads(json.dumps(node_meta, default=str))
+
+            system_msg_id = None
+            if branch.system:
+                sys_dict = branch.system.to_dict(mode="db")
+                system_msg_id = sys_dict["id"]
+                await db.insert_message(sys_dict)
+
+            br_model: str | None = None
+            br_provider: str | None = None
+            try:
+                from lionagi.state import provenance as _provenance
+
+                ep_cfg = branch.chat_model.endpoint.config
+                br_provider = getattr(ep_cfg, "provider", None)
+                br_model_raw = (ep_cfg.kwargs or {}).get("model")
+                br_model = _provenance.resolve_model_spec(br_provider, br_model_raw)
+            except Exception as _prov_exc:
+                _log_orch.debug("branch provenance lookup failed for %s: %s", branch_id, _prov_exc)
+
+            await db.create_branch(
+                {
+                    "id": branch_id,
+                    "created_at": branch_dict["created_at"],
+                    "node_metadata": node_meta,
+                    "user": branch_dict.get("user"),
+                    "name": branch_dict.get("name"),
+                    "session_id": session_id,
+                    "progression_id": branch_prog_id,
+                    "system_msg_id": system_msg_id,
+                    "model": br_model,
+                    "provider": br_provider,
+                    "agent_name": branch_dict.get("name"),
+                }
+            )
+            initialized["done"] = True
+
+    async def _on_message(msg):
+        try:
+            await _ensure_branch_row()
+            msg_dict = msg.to_dict(mode="db")
+            msg_id = msg_dict["id"]
+            await db.insert_message(msg_dict)
+            await db.append_to_progression(branch_prog_id, msg_id)
+            await db.append_to_progression(session_prog_id, msg_id)
+            await db.touch_session_activity(session_id, at=msg_dict.get("created_at"))
+            if msg_dict.get("role") == "system":
+                await db.update_branch(branch_id, system_msg_id=msg_id)
+        except Exception as exc:
+            _log_orch.warning(
+                "live persist write failed for branch %s: %s",
+                branch_id,
+                exc,
+                exc_info=True,
+            )
+
+    from lionagi.hooks import route_message_persistence
+
+    handler = route_message_persistence(ctx["session"], branch, _on_message)
+    ctx["hooks"].append((branch, handler))
 
 
 async def start_live_persist(
@@ -540,7 +718,7 @@ async def stop_live_persist(
 ) -> str:
     ctx = env._live_persist
     extras = getattr(env, "_finalize_extras", None)
-    final_status = await teardown_orchestration_persist(
+    final_status = await teardown_persist(
         ctx,
         status=status,
         exception=exception,
